move some enums to library (e.g. `HeaderVersion`)
